move config add cloud
adding dependency for google cloud storage
Allowing you to pull from a bucket into the
detector.  MOdifying setup_globals to be able
to handle the data source based on the
json cconfiguration.  Which had to be moved
into the src folder so that it would be included
in the install